+++ src/components/MyPosts/MyPosts.js
@@ -3,18 +3,17 @@ import Header from "../Header/Header";
 import { PageContainer, ContentContainer } from "../_shared/PageContainer";
 import { PageTitle } from "../_shared/PageTitle";
 import HashtagBox from "../HashtagBox/HashtagBox";
-import Post from "../_shared/Post";
+import Post from "../Post/Post";
 import { getUserPosts } from "../../API/requests";
 import routes from "../../routes/routes";
-import PagePostsContext from '../../contexts/PagePostsContext';
+import PagePostsContext from "../../contexts/PagePostsContext";
 import UserContext from "../../contexts/UserContext";
 import { useContext, useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
 
-
 export default function MyPosts() {
 	const { loggedUser } = useContext(UserContext);
-    const { pagePosts, setPagePosts } = useContext(PagePostsContext);
+	const { pagePosts, setPagePosts } = useContext(PagePostsContext);
 	const { user, token } = loggedUser;
 	const history = useHistory();
 	const [loading, setLoading] = useState(true);
@@ -26,7 +25,7 @@ export default function MyPosts() {
 
 		getUserPosts({ id: user.id, token: token })
 			.then((response) => {
-                setPagePosts(response.data.posts);
+				setPagePosts(response.data.posts);
 
 				setLoading(false);
 			})
@@ -38,23 +37,23 @@ export default function MyPosts() {
 
 	return (
 		<PageContainer>
-			{loading ?
-				<CircleLoader customStyle={{ height: '50vh' }} />
-				:
+			{loading ? (
+				<CircleLoader customStyle={{ height: "50vh" }} />
+			) : (
 				<>
 					<Header />
 
 					<ContentContainer>
 						<PageTitle>my posts</PageTitle>
 
-						{pagePosts.map(
-							(postData, index) => <Post postData={postData} key={index} />
-						)}
+						{pagePosts.map((postData, index) => (
+							<Post postData={postData} key={index} />
+						))}
 					</ContentContainer>
 
 					<HashtagBox />
 				</>
-			}
+			)}
 		</PageContainer>
 	);
 }
